refactor: Refactored and split up modules
