+++ docs/game.js
@@ -1,15 +1,11 @@
-import Player from './player';
+/* eslint-disable import/no-cycle */
+/* eslint-disable max-len */
 /* eslint-disable no-undef */
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-use-before-define */
-
-const gameBoard = (() => {
-  const board = [1, 2, 3, 4, 5, 6, 7, 8, 9];
-  return {
-    board,
-  };
-})();
-
+import Player from './player';
+import domManipulation from './dom';
+import gameBoard from './gameBoard';
 
 const gameLogic = (() => {
   let player1;
@@ -36,101 +32,39 @@ const gameLogic = (() => {
     return winMessage;
   };
 
-  const winner = (player, draw = false) => {
-    const message = document.getElementById('message');
-    while (message.firstChild) {
-      message.removeChild(message.firstChild);
-    }
-    const messageContainer = document.createElement('div');
-    const messageHeader = document.createElement('h3');
-    const messageText = document.createElement('p');
-    const winMessage = winnerMessage(draw, player);
-    messageHeader.innerHTML = winMessage;
-    messageText.innerHTML = 'would you like to play again?';
-    messageContainer.append(messageHeader);
-    messageContainer.append(messageText);
-    const buttonDiv = document.createElement('div');
-    buttonDiv.innerHTML = '<button id="reset_button">Play again!</button>';
-    message.append(messageContainer);
-    message.append(buttonDiv);
-    message.attributes.class.value = 'message_section';
-    game.attributes.class.value = 'none';
-    game_info.attributes.class.value = 'none';
-    document.querySelector('button').addEventListener('click', reset);
-  };
-
   const winCheck = () => {
     if (winningMoves.some((win) => win.every((r) => currentPlayer.comb.includes(r)))) {
-      winner(currentPlayer);
+      domManipulation.winner(currentPlayer);
     } else if (turns === 9 && !winningMoves.some((win) => win.every((r) => currentPlayer.comb.includes(r)))) {
-      winner('N/A', true);
+      domManipulation.winner('N/A', true);
     }
   };
 
-  const playerMove = (space, p1, p2) => {
-    if (p1.comb.includes(space - 1) || p2.comb.includes(space - 1)) {
-      // eslint-disable-next-line no-alert
-
-      // playerMove(space, player1, player2);
+  const playerMove = (space) => {
+    if (player1.comb.includes(space - 1) || player2.comb.includes(space - 1)) {
       return false;
     }
     return true;
   };
 
-  const displayTurn = () => {
-    const gameInfo = document.getElementById('game_info');
-    while (gameInfo.firstChild) {
-      gameInfo.removeChild(gameInfo.firstChild);
-    }
-    const turn = document.createElement('div');
-    turn.innerHTML = `<h2>${currentPlayer.getName()} it's your turn!</h2>`;
-    gameInfo.append(turn);
+  // const checkTurns = () => {
+  //   // eslint-disable-next-line max-len
+  //   if (turns === 9 && !winningMoves.some((win) => win.every((r) => currentPlayer.comb.includes(r)))) {
+  //     winner('N/A', true);
+  //   }
+  // };
+
+  const switchPlayer = () => {
+    // console.log(currentPlayer.getName());
+    currentPlayer = currentPlayer === player1 ? player2 : player1;
+    return currentPlayer;
   };
 
-  const checkTurns = () => {
-    // eslint-disable-next-line max-len
-    if (turns === 9 && !winningMoves.some((win) => win.every((r) => currentPlayer.comb.includes(r)))) {
-      winner('N/A', true);
-    }
+  const incrementmoves = (space) => {
+    turns += 1;
+    gameBoard.board[space - 1] = currentPlayer.getSymbol();
+    currentPlayer.comb.push(space - 1);
   };
-
-  function checkMove(check, space, boardSpaceText) {
-    if (check) {
-      gameBoard.board[space - 1] = currentPlayer.getSymbol();
-      boardSpaceText.innerHTML = currentPlayer.getSymbol();
-      currentPlayer.comb.push(space - 1);
-      winCheck();
-      //checkTurns();
-      currentPlayer = currentPlayer === player1 ? player2 : player1;
-      displayTurn();
-    } else {
-      alert('this position has already been taken chose another');
-    }
-    console.log(boardSpaceText.innerHTML);
-    return boardSpaceText;
-  }
-
-  function render() {
-    const game = document.getElementById('game');
-    while (game.firstChild) {
-      game.removeChild(game.firstChild);
-    }
-    displayTurn();
-    const boardArr = gameBoard.board;
-    boardArr.forEach(space => {
-      const boardSpace = document.createElement('a');
-      const boardSpaceText = document.createElement('p');
-      boardSpaceText.innerHTML = space;
-      // eslint-disable-next-line func-names
-      boardSpace.onclick = () => {
-        turns += 1;
-        const check = playerMove(space, player1, player2);
-        checkMove(check, space, boardSpaceText);
-      };
-      boardSpace.append(boardSpaceText);
-      game.append(boardSpace);
-    });
-  }
 
   const switchForm = () => {
     const newPlayerForm = document.getElementById('player');
@@ -152,7 +86,7 @@ const gameLogic = (() => {
     game_info.attributes.class.value = 'game_info';
     message.attributes.class.value = 'none';
     currentPlayer = player1;
-    render();
+    domManipulation.render(currentPlayer, player1, player2);
   }
 
   function gameInit(e) {
@@ -163,24 +97,21 @@ const gameLogic = (() => {
     player2 = Player(tempplayer2, 'O');
     currentPlayer = player1;
     switchForm();
-    render();
+    domManipulation.render(currentPlayer, player1, player2);
   }
 
 
   return {
     winnerMessage,
-    winner,
     winCheck,
     playerMove,
-    displayTurn,
-    checkTurns,
-    render,
     switchForm,
     reset,
     gameInit,
-    winningMoves,
+    incrementmoves,
+    switchPlayer,
   };
 })();
 
 
-export { gameBoard, gameLogic };
+export default gameLogic;
