+++ docs/gameBoard.js
@@ -1,0 +1,8 @@
+const gameBoard = (() => {
+  const board = [1, 2, 3, 4, 5, 6, 7, 8, 9];
+  return {
+    board,
+  };
+})();
+
+export default gameBoard;
