+++ docs/index.js
@@ -1,3 +1,3 @@
-import { gameLogic } from './game';
+import gameLogic from './game';
 
 document.getElementById('newPlayerForm').addEventListener('submit', gameLogic.gameInit);
